% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{adjusted_subset_vs_whole_test}
\alias{adjusted_subset_vs_whole_test}
\title{Perform multiple hypothesis tests on subset vs whole (without subset)}
\usage{
adjusted_subset_vs_whole_test(
  input_df,
  grouping_colname,
  metric_colname,
  list_of_subset_groups,
  test_to_use
)
}
\arguments{
\item{input_df}{A dataframe that contains: a column with a grouping variable
from which one can extract a subset of the data and a column containing
a metric that the user wishes to compare between subset and rest of dataset.}

\item{list_of_subset_groups}{A list of string literals that specify
subgroups that exist in grouping_column}

\item{test_to_use}{A string literal which match the follow: "student", "welch",
"mann_whitney". Specifies a test to use.}

\item{grouping_column}{The column that contains the grouping variable, such
as patient subtype, etc.}

\item{metric_column}{The column that contains a metric the user would like
to compare between two subsets of the dataset.}
}
\value{
A dataframe of test results, with adjusted p-values.
}
\description{
Perform multiple hypothesis tests on subset vs whole (without subset)
}
