% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{append_principal_components}
\alias{append_principal_components}
\title{Inner join principal components to a dataframe}
\usage{
append_principal_components(
  df_to_append,
  principal_component_df,
  colname_to_append_by
)
}
\arguments{
\item{df_to_append}{A dataframe to append principal components}

\item{principal_component_df}{A dataframe with principal components}

\item{colname_to_append_by}{A string specifying a column that is present in both df_to_append and principal_component_df}
}
\value{
A single dataframe that has principal components appended to df_to_append
}
\description{
Inner join principal components to a dataframe
}
\examples{
pc_1_2 <- perform_and_extract_pca(rnaseq_counts, gene_id, 2)
df <- some_data_frame
df_with_pc_1_2 <- append_principal_components(
     df,
     pc_1_2,
     "some_colname"
)
}
