% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression_analysis.r
\name{arrange_df_by_norm_factors}
\alias{arrange_df_by_norm_factors}
\title{Arrange raw counts by the order of specified factors for differential expression analysis}
\usage{
arrange_df_by_norm_factors(
  raw_counts,
  df_with_factor,
  ref_factor,
  compare_factor,
  factor_colname = "cluster",
  id_colname = "patient_id"
)
}
\arguments{
\item{raw_counts}{A dataframe of raw counts for use in expression analysis,
with sample/patient IDs as columns}

\item{df_with_factor}{A dataframe that contains a column of factors to
identify groups for expression analysis}

\item{ref_factor}{A string literal for the reference factor found in factor}

\item{compare_factor}{A stirng literal for the factor contrasted against
colname the reference factor, found in factor_colname}

\item{factor_colname}{A string literal specifying a column in df_with_factor}

\item{id_colname}{A string literal specifying a column that contains
sample/patient IDs}
}
\value{
A dataframe of raw counts with columns ordered by a specified set of factors
}
\description{
Arrange raw counts by the order of specified factors for differential expression analysis
}
\examples{
raw_counts <- some_rnaseq_counts
df <- some_df_with_reference_and_compare_factors
arranged_df <- arrange_df_by_norm_factors(
     raw_counts = raw_counts,
     df_with_factor = df,
     factor_colname = "factor_column_name",
     ref_factor =  ref_factor,
     compare_factor = compare_factor,
     id_colname = "id_column_name"
)
}
