% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.r
\name{assign_maxstat_cutoff}
\alias{assign_maxstat_cutoff}
\title{Find maxstat survival cutoff}
\usage{
assign_maxstat_cutoff(
  input_df,
  metric_colname,
  censor_colname,
  survival_time_colname
)
}
\arguments{
\item{input_df}{The dataframe containing the metric}

\item{metric_colname}{A string literal that refers to a column in input_df}

\item{censor_colname}{A string literal referring to a column of censors}

\item{survival_time_colname}{A string literal referring to a column for survival time}
}
\value{
input_df but with an extra column detailing median status
}
\description{
Adds a column that determines if a metric is above or below the
maxstat cutoff.
}
