% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.r
\name{assign_median_cutoff}
\alias{assign_median_cutoff}
\title{Find median cutoff}
\usage{
assign_median_cutoff(input_df, metric_colname)
}
\arguments{
\item{input_df}{The dataframe containing the metric}

\item{metric_colname}{A string literal that refers to a column in input_df}
}
\value{
input_df but with an extra column detailing median status
}
\description{
Adds a column that determines if a sample is above or below median
in a given metric
}
