% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{create_list_of_subset_comparisons}
\alias{create_list_of_subset_comparisons}
\title{Create multiple dataframes with subset vs whole(without subset)}
\usage{
create_list_of_subset_comparisons(
  input_df,
  grouping_colname,
  metric_colname,
  list_of_subset_groups
)
}
\arguments{
\item{input_df}{A dataframe that contains: a column with a grouping variable
from which one can extract a subset of the data and a column containing
a metric that the user wishes to compare between subset and rest of dataset.}

\item{list_of_subset_groups}{A list of string literals specifying subgroups
of the grouping column.}

\item{grouping_column}{The column that contains the grouping variable, such
as patient subtype, etc.}

\item{metric_column}{The column that contains a metric the user would like
to compare between two subsets of the dataset.}
}
\value{
A list of adataframes of the format of define_subset_against_whole()
}
\description{
This creates a list of multiple different dataframes, consisting of
subset vs whole (without subset) columns.
}
