% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{decide_which_hypothesis_test}
\alias{decide_which_hypothesis_test}
\title{Determine which hypothesis test is appropriate for distributions}
\usage{
decide_which_hypothesis_test(
  input_df,
  grouping_colname,
  metric_colname,
  subset
)
}
\arguments{
\item{input_df}{A dataframe that contains: a column with a grouping variable
from which one can extract a subset of the data and a column containing
a metric that the user wishes to compare between subset and rest of dataset.}

\item{subset}{A string literal that specifies a subgroup in the grouping
column}

\item{grouping_column}{The column that contains the grouping variable, such
as patient subtype, etc.}

\item{metric_column}{The column that contains a metric the user would like
to compare between two subsets of the dataset.}
}
\value{
A string literal of: "mann_whitney", "welch", "student" or "other".
These represent the appropriate hypothesis test to use given the
distributions of the subset and whole (without subset) portions of the
dataset. The "other" represents a case where all the tests fail and
an alternative test should be used.
}
\description{
There are three hypothesis tests that are considered for this function:
Student t-test (assumption of variance is equal), Welch's t-test
(assumotion of unequal variances) when distribution of residuals
are normal. When normality is not achieved, this function
will assess if the Mann-Whitney U-test is appropriate by testing
if distributions are equal via. the Kolomogrov-Smirnov test.
}
