% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{define_subset_against_whole}
\alias{define_subset_against_whole}
\title{Create a dataframe with grouping variables subset vs whole (without subset)}
\usage{
define_subset_against_whole(
  input_df,
  grouping_colname,
  subset_group,
  metric_colname
)
}
\arguments{
\item{input_df}{A dataframe that contains: a column with a grouping variable
from which one can extract a subset of the data and a column containing
a metric that the user wishes to compare between subset and rest of dataset.}

\item{subset_group}{A string literal present in grouping_column that defines
the subset subgroup the user is interested in.}

\item{grouping_column}{The column that contains the grouping variable, such
as patient subtype, etc.}

\item{metric_column}{The column that contains a metric the user would like
to compare between two subsets of the dataset.}
}
\value{
A dataframe with 3 columns: grouping_column, metric_column and group,
which is a column that defines which part of the data belongs to the subset
and the rest of the dataset.
}
\description{
The purpose of this function is to create a dataframe where one wants to
compare a subset of a dataset against the rest of the dataset. The main
reason to do this is to maintain the assumption of indepdence in
many inferrential hypohtesis tests. This would allow the user
to see how much a metric changes in the subset, relative to the
rest of the data in the dataset.
}
