% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{define_subset_comparisons}
\alias{define_subset_comparisons}
\title{Define groups: Subset of dataset and Dataset without subset}
\usage{
define_subset_comparisons(
  input_df,
  subset_group,
  reference_group,
  identifier_column,
  sample_column
)
}
\arguments{
\item{input_df}{A dataframe that contains samples the user wishes to compare.
This dataframe should have a grouping for both the subset and the whole
dataset.}

\item{subset_group}{A string literal that defines a subset of samples the
user wishes to compare to the rest of the dataset.}

\item{reference_group}{A string literal that defines samples belonging to the
whole dataset.}

\item{identifier_column}{A column that contains the grouping variables
subset_group and reference_group.}

\item{sample_column}{A column that contains the sample IDs for this dataset.}
}
\value{
A dataframe with a column that specifies a comparison between
the outlined subset and the rest of the dataset.
}
\description{
This function separates samples into two groups. The first, a subset of
samples (such as a group of patients belonging to a particular substype)
and the second, the rest of the dataset without samples of this subset.
This function assumes that there is a column specifying sample groupings,
and that there is both a grouping for the subset the user is interested in
as well as a grouping for all samples in the dataset.
This function is used to set hypotheses test downstream, if the user
wishes to compare a subset against an "All" group. This is particularly
useful especially when samples can inhabit multiple different subtypes
for example a cancer patient can harbour multiple mutations. If
the user is interested to see if the expression of a gene is enriched
in a particular subset of patients, this can serve as a statistical
basis in testing if this is actually so.
The main purppose of this function is maintain the assumption of
independence that many inferrential hypothesis tests require.
}
\examples{
# Example for data preparation prior to using the function
data(iris)
# Defining a grouping that encapsulates the entire dataset
all_group <- iris \%>\% dplyr::mutate(group = "all")

# Defining a subset of the dataset
sestosa_group <- iris \%>\%
     dplyr::filter(Species \%in\% "sestosa") \%>\%
     dplyr::mutate(group = "sestosa")

# Combining dataframes, this can then be used in the function.
# The group column will be identifier_column.
whole_and_subset <- rbind(all_group, sestosa_group)
}
