% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression_analysis.r
\name{ensembl_to_entrez_and_symbol}
\alias{ensembl_to_entrez_and_symbol}
\title{Convert ENSEMBL IDs into gene symbols and entrez-IDs}
\usage{
ensembl_to_entrez_and_symbol(deg_df, id_colname = "ensembl_id")
}
\arguments{
\item{deg_df}{A dataframe that is a result of the topTags edgeR function.
Can be a dataframe returned from find_degs function.}

\item{id_colname}{A string literal that specifies the column in the dataframe
reserved for ENSEMBL IDs. By default, assumed to be "ensembl_id".}
}
\value{
deg_df with two new columns, one for gene symbols, another for
entrez IDs
}
\description{
Convert ENSEMBL IDs into gene symbols and entrez-IDs
}
