% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_utils.r
\name{extract_clusters_from_kmeans}
\alias{extract_clusters_from_kmeans}
\title{Extract clusters from a kmeans_result object}
\usage{
extract_clusters_from_kmeans(kmeans_cluster_result)
}
\arguments{
\item{kmeans_cluster_result}{A result from the stats::kmeans() function}
}
\value{
A dataframe with columns "patient_id" and "cluster", where cluster
specifies clustering found via kmeans clustering.
}
\description{
This function takes a kmeans_result object as input, and will extract
the cluster classifcation for each sample as a dataframe.
The assumption is that sample or patient IDs are stored as rownames
prior to clustering, for example at dimensionality reduction step.
}
