% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{extract_gene_expression}
\alias{extract_gene_expression}
\title{Extract gene expression from normalize_counts() result}
\usage{
extract_gene_expression(normalized_counts, gene_column, gene)
}
\arguments{
\item{normalized_counts}{A dataframe from normalize_counts() where
samples are columns, genes are rwos.}

\item{gene_column}{A column in normalized_counts that specifies where
the gene symbols are stored.}

\item{gene}{A string literal representing the gene(s) of interest.}
}
\value{
A dataframe with two columns, patient_ids and log2_expression
}
\description{
This function takes a dataframe which is the result from normalize_counts(),
and pivots it such that there are now only two columns, patient_ids and
log2_expression. The input dataframe is filtered for gene(s) of interest,
and then pivoted.
}
