% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.r
\name{fit_survival_curve_by_factor}
\alias{fit_survival_curve_by_factor}
\title{Fit survival curves by a grouping factor}
\usage{
fit_survival_curve_by_factor(
  survival_df,
  survival_time_colname,
  survival_censor_colname,
  factor_colname
)
}
\arguments{
\item{survival_df}{A dataframe that contains survival information}

\item{survival_time_colname}{A string literal that refers to a column
in survival_df that outlines patient survival time}

\item{survival_censor_colname}{A string literal that refers to a column
in survial_df that outlines patient survival censoring}

\item{factor_colname}{A string literal that refers to a column in
survival_df that contains a grouping variable used to categorize patients.
Survival between these groups will be measured.}
}
\value{
An object with fitted survival curves
}
\description{
This function takes a dataframe that contains survival data, specifically
columns for survival time and survival censor, as well as a grouping
variable that is detailed in a column, factor_colname.
}
