% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize raw counts using edgeR}
\usage{
normalize_counts(input_df, ids_to_use, filter_method = "liberal")
}
\arguments{
\item{input_df}{A dataframe of raw RNA-seq counts, samples are columns.}

\item{ids_to_use}{Patient IDs used to select a subset of patients.}

\item{filter_method}{A string literal of either "strict" or "liberal",
liberal will take genes with rowSums > 50, strict will use edgeR's
filterByExpr which takes a cpm of 10 (by default) in a significant
number of samples.}
}
\value{
A dataframe of log2 transformed TMM normalized counts,
where patients are columns, genes are rows
}
\description{
This function takes a dataframe of raw RNA-seq counts (samples as columns),
and normalizes it using edgeR's TMM method.
It returns a dataframe of log2 transofrmed normalized counts.
This function will select a subset of patients.
This functions offers the option of liberal or strict
count filtering. Liberal filtering sees if counts per gene
are greater than 50. The strict method utilizes edgeR's
filterByExpr function and asserts that a cpm of 10 (default value)
is achieved in a significant amount of samples.
}
