% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{perform_and_extract_pca}
\alias{perform_and_extract_pca}
\title{Perform PCA and extract specified amount of principal components}
\usage{
perform_and_extract_pca(
  normalized_counts,
  gene_id_colname,
  number_of_pc_to_extract
)
}
\arguments{
\item{normalized_counts}{Normalized RNA-seq counts, where columns are samples}

\item{gene_id_colname}{The column name in normalized_counts for gene ids}

\item{number_of_pc}{The number of principal components to be extracted from prcomp_result}
}
\value{
A dataframe containing the specified amount of principal components, with PCs as columns
}
\description{
Perform PCA and extract specified amount of principal components
}
\examples{
# Extract first 2 principal components from normalized counts
pc_1_and_2 <- perform_and_extract_pca(some_ranseq_counts, gene_id_colname, 2)
}
