% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{perform_and_plot_kruskal_test}
\alias{perform_and_plot_kruskal_test}
\title{Perform and plot a Kruskal-Wallis test}
\usage{
perform_and_plot_kruskal_test(df, expression, group)
}
\arguments{
\item{df}{A dataframe that contains the grouping variable and the
metric to compare across groups.}

\item{expression}{A string literal for the column representing gene
expression.}

\item{group}{A string literal for the column representing the
grouping variable.}
}
\value{
A boxplot comparing expression across groups. Plots significance
baed on results of Dunn's test. The p-value obtained from the Kruskal Wallis
along with number of samples compared and degrees of freedom will be reported
at the top of the plot.
}
\description{
Given a grouping variable, this function will perform a Kruskal-Wallis
test, along with Dunn's test as a post-hoc test.
}
