% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.r
\name{perform_gsea}
\alias{perform_gsea}
\title{Perform Gene Set Enrichment Analysis on differentially expressed genes}
\usage{
perform_gsea(df_of_degs, pathway_annotation)
}
\arguments{
\item{df_of_degs}{A dataframe of differentially expressed genes.
Must contain a column for entrez geneIDs and logFC. Anticipated to
be a result of edgeR's topTags funcion.}

\item{pathway_annotation}{Must be a string literal matching one of
"kegg", "go" or "reactome". Specifies which pathway annotation to
use for pathway analysis.}
}
\value{
A gseaResult object
}
\description{
Utilize clusterProfiler's GSEA algorithim to perform pathway analysis
on a dataframe of differentially expressed genes. It is expected
that the input dataframe of differentially expressed genes has
a column for entrez gene IDs and a column for log fold change called
"logFC".
The user can choose 3 different pathway annotation databases:
KEGG, Gene Ontology and Reactome. For Gene Ontology, all the
databases are used.
}
