% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_utils.r
\name{perform_kmeans_and_extract_clusters}
\alias{perform_kmeans_and_extract_clusters}
\title{Perform kmeans clustering and extract cluster information}
\usage{
perform_kmeans_and_extract_clusters(
  random_seed,
  input_df,
  centers,
  nstart = 25
)
}
\arguments{
\item{random_seed}{An unsigned number that specifies a random seed.
This should bethe same random seed used in calculating the kmeans
clustering used in determining the optimal amount of clusters.}

\item{input_df}{A dataframe with variables as columns, samples as rows.
It is assumed the the rows are patient IDs.}

\item{centers}{The number of clusters to use in kmeans clustering.}

\item{nstart}{An unsigned number, defaults to 25. Determines how many
random sets should be chosen. ?kmeans for more information.}
}
\value{
A dataframe with columns of patient IDs and cluster information
}
\description{
This function performs kmeans clustering on a dataframe, where variables
are columns and samples are rows. It takes an argument, centers, which
should be the ideal amount of clustering, found through the
plot_optimal_cluster_tests function.
}
