% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{perform_pca_on_counts}
\alias{perform_pca_on_counts}
\title{Perform Principal Component analysis on normalized RNA-seq counts}
\usage{
perform_pca_on_counts(normalized_counts, gene_id_colname)
}
\arguments{
\item{normalized_counts}{Normalized RNA-seq counts, where columns are samples}

\item{gene_id_colname}{The column name in normalized_counts for gene ids}
}
\value{
A prcomp object of centred and scaled principal components, with patients as rows and genes as columns
}
\description{
Perform Principal Component analysis on normalized RNA-seq counts
}
\examples{
# Do not run
# rna_seq_counts <- some_data_frame
# pca_result <- perform_pca_on_counts(rna_seq_counts)
}
