% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{perform_subset_wilcox}
\alias{perform_subset_wilcox}
\title{Compare a subset against the rest of dataset using Mann-Whitney U Tests.}
\usage{
perform_subset_wilcox(input_df, identifier_colname, gene)
}
\description{
Compare a subset against the rest of dataset using Mann-Whitney U Tests.
}
