% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{plot_each_subset_comparison}
\alias{plot_each_subset_comparison}
\title{Create a plot of multiple subset vs rest comparisons}
\usage{
plot_each_subset_comparison(
  grouping_colname,
  metric_colname,
  list_of_subset_groups,
  input_df
)
}
\arguments{
\item{list_of_subset_groups}{A list of string literals specifying
subgroups that exist in grouping_column}

\item{input_df}{A dataframe that contains: a column with a grouping variable
from which one can extract a subset of the data and a column containing
a metric that the user wishes to compare between subset and rest of dataset.}

\item{grouping_column}{The column that contains the grouping variable, such
as patient subtype, etc.}

\item{metric_column}{The column that contains a metric the user would like
to compare between two subsets of the dataset.}
}
\value{
A plot of multiple plots
}
\description{
Create a plot of multiple subset vs rest comparisons
}
