% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_utils.r
\name{plot_optimal_cluster_tests}
\alias{plot_optimal_cluster_tests}
\title{Plot elbow, gap statistic and silhouette plots to determine optimal amount
of clusters for kmeans clustering}
\usage{
plot_optimal_cluster_tests(input_df)
}
\arguments{
\item{input_df}{A dataframe with variables as columns, samples as rows}
}
\value{
Prints 3 plots to test optimal cluster amounts for k-means clustering
}
\description{
This function will print plots for each test to help the user determine
how many clusters is appropriate for their analysis.
}
