% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{plot_subset_comparisons}
\alias{plot_subset_comparisons}
\title{Create a plot comparing subset and rest of data}
\usage{
plot_subset_comparisons(input_df, grouping_colname, metric_colname)
}
\arguments{
\item{input_df}{A dataframe that contains: a column with a grouping variable
from which one can extract a subset of the data and a column containing
a metric that the user wishes to compare between subset and rest of dataset.}

\item{grouping_column}{The column that contains the grouping variable, such
as patient subtype, etc.}

\item{metric_column}{The column that contains a metric the user would like
to compare between two subsets of the dataset.}
}
\value{
A plot
}
\description{
Create a plot comparing subset and rest of data
}
