% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.r
\name{save_dotplot}
\alias{save_dotplot}
\title{Automatically save dotplot of pathway analysis}
\usage{
save_dotplot(pathway_analysis, title, path = "results/")
}
\arguments{
\item{pathway_analysis}{A gseaResult or enrichResult object}

\item{title}{Title of the file to be saved}

\item{path}{Path of the file destination. Defaults to "results/"}
}
\value{
A 8in (width) x 6in (height) pdf file. Dotplot shows top 10 pathways.
}
\description{
Automatically save dotplot of pathway analysis
}
