% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.r
\name{save_ridgeplot}
\alias{save_ridgeplot}
\title{Automatically save ridgeplot of pathway analysis}
\usage{
save_ridgeplot(
  pathway_analysis,
  title,
  path = "results/",
  number_of_pathways = 15
)
}
\arguments{
\item{pathway_analysis}{A gseaResult object}

\item{title}{Title of the file to be saved}

\item{path}{Path of the file destination. Defaults to "results/"}

\item{number_of_pathways}{An unsigned number that specifies how many
pathways to show in the ridgeplot. Defaults to 15.}
}
\value{
A ridgeplot of 8in (width) x 6in (height) pdf file. Ridgeplot shows
specified number of pathways. Defaults to 15.
}
\description{
NOTE: This function does not work with ORA analyses.
}
