% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.r
\name{select_ids_from_subtype}
\alias{select_ids_from_subtype}
\title{Select Patient/Sample IDs based on subtype}
\usage{
select_ids_from_subtype(input_df, id_column, subtype_column, subtype_to_filter)
}
\arguments{
\item{input_df}{A dataframe which contains a column specifying
patient subtypes}

\item{id_column}{The column where sample/patient IDs are stored}

\item{subtype_column}{The column where patient subtypes are stored}

\item{subtype_to_filter}{A stirng literal represting a subtype
present in subtype_column.}
}
\value{
A vector of characters specifying IDs for a respective subtype
}
\description{
Select Patient/Sample IDs based on subtype
}
