% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{subset_vs_whole_mann_whitney}
\alias{subset_vs_whole_mann_whitney}
\title{Compare a subset against the rest of dataset using Mann-Whitney U Tests.}
\usage{
subset_vs_whole_mann_whitney(
  input_df,
  grouping_colname = "group",
  metric_colname
)
}
\arguments{
\item{input_df}{A dataframe that contains: a column with a grouping variable
from which one can extract a subset of the data and a column containing
a metric that the user wishes to compare between subset and rest of dataset.}

\item{grouping_column}{The column that contains the grouping variable, such
as patient subtype, etc.}

\item{metric_column}{The column that contains a metric the user would like
to compare between two subsets of the dataset.}
}
\value{
A dataframe detailing results of the test.
}
\description{
Compare a subset against the rest of dataset using Mann-Whitney U Tests.
}
