% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{subset_vs_whole_t_test}
\alias{subset_vs_whole_t_test}
\title{Perform a t-test on a subset against the rest of a dataset}
\usage{
subset_vs_whole_t_test(
  input_df,
  grouping_colname = "group",
  metric_colanme,
  equal_variance = FALSE
)
}
\arguments{
\item{input_df}{A dataframe that contains: a column with a grouping variable
from which one can extract a subset of the data and a column containing
a metric that the user wishes to compare between subset and rest of dataset.}

\item{equal_variance}{A boolean value stating whether or not the
assumption of equal variance is TRUE or FALSE. If TRUE, a student's t-test
will be performed. If FALSE, then Welch's t-test instead. The default is
FALSE, as R assumes that variance is unequal by default.}

\item{grouping_column}{The column that contains the grouping variable, such
as patient subtype, etc.}

\item{metric_column}{The column that contains a metric the user would like
to compare between two subsets of the dataset.}
}
\value{
A dataframe detailing results of the test.
}
\description{
Perform a t-test on a subset against the rest of a dataset
}
