% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{verify_equality_of_variances}
\alias{verify_equality_of_variances}
\title{Verify equality of variances in samples grouped by a variable}
\usage{
verify_equality_of_variances(input_df, grouping_colname, metric_colname)
}
\arguments{
\item{input_df}{A dataframe that contains: a column with a grouping variable
from which one can extract a subset of the data and a column containing
a metric that the user wishes to compare between subset and rest of dataset.}

\item{grouping_column}{The column that contains the grouping variable, such
as patient subtype, etc.}

\item{metric_column}{The column that contains a metric the user would like
to compare between two subsets of the dataset.}
}
\value{
A boolean value, TRUE if the p-value obtained from Levene's test is
greater than 0.05 and FALSE if the p-value is less tha 0.05.
}
\description{
This function uses Levene's test to determine if variance is equal
amongst all subgroups.
}
