% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_analysis.r
\name{verify_normality}
\alias{verify_normality}
\title{Verify if distribution amongst grouping variables is normal}
\usage{
verify_normality(input_df, grouping_colname, metric_colname, subset = "subset")
}
\arguments{
\item{input_df}{A dataframe that contains: a column with a grouping variable
from which one can extract a subset of the data and a column containing
a metric that the user wishes to compare between subset and rest of dataset.}

\item{subset}{A string literal that specifies a subgroup in the grouping
column}

\item{grouping_column}{The column that contains the grouping variable, such
as patient subtype, etc.}

\item{metric_column}{The column that contains a metric the user would like
to compare between two subsets of the dataset.}
}
\value{
A boolean value, TRUE if the p-value obtained from Shapiro-Wilk's
test is greater than 0.05 and FALSE if the p-value is less tha 0.05.
}
\description{
This function determines if the distribution of a particular metric
is normal, across grouping variables, using a Shapiro-Wilk test. It
is expected that this function will be used to assess which test is
valid for the subset vs whole (without subset) comparisons.
}
